refactor: Establish FSD 'pages' hierarchy
