+++ backend/bin/main/application.properties
@@ -1,21 +1,22 @@
 spring.application.name=icebreaker
 
-server.port=4000
-
+server.address=0.0.0.0
+server.port=8080
+server.time-zone=Asia/Seoul
 
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://127.0.0.1:3306/ICEBREAKER?serverTimezone=UTC&characterEncoding=UTF-8
-spring.datasource.username=been1
-spring.datasource.password=P!ssw0rd
+spring.datasource.url=${DATABASE_URL}
+spring.datasource.username=${DATABASE_USERNAME}
+spring.datasource.password=${DATABASE_PASSWORD}
 
 spring.servlet.multipart.max-file-size=100MB
 spring.servlet.multipart.max-request-size=110MB
+spring.jpa.properties.hibernate.jdbc.time_zone=UTC
 
-# Gmail 
+# Gmail SMTP 설정
 spring.mail.host = smtp.gmail.com
 spring.mail.port = 587
-spring.mail.username = hufsicebreaker@gmail.com
-spring.mail.password = kcub ttjq dapt gcvg
+spring.mail.username = ${MAIL_USERNAME}
+spring.mail.password = ${MAIL_PASSWORD}
 spring.mail.properties.mail.smtp.auth = true
 spring.mail.properties.mail.smtp.starttls.enable = true
-
