+++ frontend/src/components/NavBar.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { Link, NavLink, useNavigate } from 'react-router-dom';
-import '../layouts/css/Main.bar/NavBar.css';
+import '../pages/css/Main.bar/NavBar.css';
 import LoginForm from '../Modals/LoginForm';
 import SignUpinfoForm from '../Modals/Signupinfo';
 import MypageForm from '../Modals/Mypage';
