+++ frontend/src/pages/css/Main.bar/MainPage.css
